Important Bug Patch - resolved concurrency issues and added guardrails to ensure that trial_ids were unique
